+++ src/test/java/ch/sbb/polarion/extension/diff_tool/rest/model/diff/WorkItemsDiffTest.java
@@ -1,6 +1,7 @@
 package ch.sbb.polarion.extension.diff_tool.rest.model.diff;
 
 import ch.sbb.polarion.extension.diff_tool.service.CalculatePairsContext;
+import com.polarion.alm.tracker.model.IModule;
 import com.polarion.alm.tracker.model.IWorkItem;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.Test;
@@ -24,7 +25,15 @@ class WorkItemsDiffTest {
         IWorkItem wiMock2 = mock(IWorkItem.class);
         when(wiMock2.getId()).thenReturn("id2");
 
-        WorkItemsPair workItemsPair = WorkItemsPair.of(Pair.of(wiMock1, wiMock2), mock(CalculatePairsContext.class));
+        CalculatePairsContext calculatePairsContext = mock(CalculatePairsContext.class);
+        IModule leftDocument = mock(IModule.class);
+        when(leftDocument.getProjectId()).thenReturn("projectId");
+        IModule rightDocument = mock(IModule.class);
+        when(rightDocument.getProjectId()).thenReturn("projectId");
+        when(calculatePairsContext.getLeftDocument()).thenReturn(leftDocument);
+        when(calculatePairsContext.getRightDocument()).thenReturn(rightDocument);
+
+        WorkItemsPair workItemsPair = WorkItemsPair.of(Pair.of(wiMock1, wiMock2), calculatePairsContext);
 
         workItemsPair.getLeftWorkItem().addField(WorkItem.Field.builder().id("id").htmlDiff("diffLeft").build());
         workItemsPair.getRightWorkItem().addField(WorkItem.Field.builder().id("id").htmlDiff("diffRight").build());